+++ src/gui/settings_io.h
@@ -44,15 +44,15 @@ struct ParsedSettings {
     // applies at the call site.
     bool    has_font_size      = false;
     double  font_size          = 11.0;
-    // Per-tab trim keys. Absence means unset. Exact source-frame doubles.
+    // Per-tab trim keys. Absence means unset. Whole int64 source frames.
     bool    has_tab_a_trim_begin = false;
-    double  tab_a_trim_begin     = 0.0;   // source frames
+    int64_t tab_a_trim_begin     = 0;     // source frames
     bool    has_tab_a_trim_end   = false;
-    double  tab_a_trim_end       = 0.0;   // source frames
+    int64_t tab_a_trim_end       = 0;     // source frames
     bool    has_tab_b_trim_begin = false;
-    double  tab_b_trim_begin     = 0.0;   // source frames
+    int64_t tab_b_trim_begin     = 0;     // source frames
     bool    has_tab_b_trim_end   = false;
-    double  tab_b_trim_end       = 0.0;   // source frames
+    int64_t tab_b_trim_end       = 0;     // source frames
     // Per-tab read-only flags. Absent → tab defaults to editable.
     bool    has_tab_a_read_only  = false;
     bool    tab_a_read_only      = false;
@@ -117,9 +117,9 @@ struct RendersettingsAuthoring {
     bool    has_active_audio_view = false;
     char    active_audio_view     = 'S';
     bool    has_trim_begin        = false;
-    double  trim_begin_frame      = 0.0;   // source frames
+    int64_t trim_begin_frame      = 0;     // source frames
     bool    has_trim_end          = false;
-    double  trim_end_frame        = 0.0;   // source frames
+    int64_t trim_end_frame        = 0;     // source frames
     bool    has_zoom_level        = false;
     int     zoom_level            = 0;
     bool    has_viewport_start    = false;
